[ADD] post create mutation
